Refactor TestGroup out of base Queue.
Move base Queue into util package, wrap TestGroup-specific functionality in TestGroupQueue
