+++ config/BUILD.bazel
@@ -12,10 +12,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pb/config:go_default_library",
+        "//util:go_default_library",
         "//util/gcs:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
         "@org_golang_google_protobuf//reflect/protoreflect:go_default_library",
     ],
